# src/main/resources/openapi-doc.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SaltyVote API",
        "description": "API documentation for the saltyvote application.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
    "paths": {
...
        "/api/v1/votes/submissions": {
            "post": {
                "tags": [
                    "vote-submission-controller"
                ],
                "operationId": "submitVote",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VoteSubmissionDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/votes/{voteId}/submissions/my": {
            "get": {
                "tags": [
                    "vote-submission-controller"
                ],
                "operationId": "getMySubmission",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteSubmissionResponseDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "vote-submission-controller"
                ],
                "operationId": "deleteMySubmission",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {}
                    }
                }
            }
        },
        "/api/v1/votes/{voteId}/results": {
            "get": {
                "tags": [
                    "vote-submission-controller"
                ],
                "operationId": "getVoteResults",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteResultsDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/votes/{voteId}/results/detailed": {
            "get": {
                "tags": [
                    "vote-submission-controller"
                ],
                "operationId": "getDetailedVoteResults",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteDetailedResultsDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VoteSubmissionDto": {
                "type": "object",
                "properties": {
                    "voteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "entries": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VoteSubmissionEntryDto"
                        }
                    }
                },
                "required": [
                    "entries",
                    "voteId"
                ]
            },
            "VoteSubmissionEntryDto": {
                "type": "object",
                "properties": {
                    "optionId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "rank": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "selected": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "optionId"
                ]
            },
...
            "VoteSubmissionResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "voteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "entries": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VoteSubmissionEntryDto"
                        }
                    }
                },
                "required": [
                    "entries",
                    "id",
                    "userId",
                    "voteId"
                ]
            },
            "VoteOptionResultDto": {
                "type": "object",
                "properties": {
                    "optionId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "label": {
                        "type": "string"
                    },
                    "voteCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "percentage": {
                        "type": "number",
                        "format": "double"
                    },
                    "averageRank": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "label",
                    "optionId",
                    "percentage",
                    "voteCount"
                ]
            },
            "VoteResultsDto": {
                "type": "object",
                "properties": {
                    "voteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "totalSubmissions": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VoteOptionResultDto"
                        }
                    }
                },
                "required": [
                    "results",
                    "title",
                    "totalSubmissions",
                    "voteId"
                ]
            },
            "IndividualVoterDto": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "email": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "submittedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "rankings": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VoterRankingDto"
                        }
                    }
                },
                "required": [
                    "email",
                    "rankings",
                    "submittedAt",
                    "userId",
                    "username"
                ]
            },
            "VoteDetailedResultsDto": {
                "type": "object",
                "properties": {
                    "voteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "voteType": {
                        "type": "string",
                        "enum": [
                            "RANKING",
                            "SIMPLE"
                        ]
                    },
                    "totalSubmissions": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VoteOptionResultDto"
                        }
                    },
                    "individualVoters": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IndividualVoterDto"
                        }
                    }
                },
                "required": [
                    "individualVoters",
                    "results",
                    "title",
                    "totalSubmissions",
                    "voteId",
                    "voteType"
                ]
            },
            "VoterRankingDto": {
                "type": "object",
                "properties": {
                    "optionId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "optionLabel": {
                        "type": "string"
                    },
                    "rank": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "selected": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "optionId",
                    "optionLabel"
                ]
            },
...
        }
    }
}